Django/requirements.txt
alembic==0.9.3
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
click==6.7
dj-database-url==0.4.2
Django==1.9.4
django-celery==3.2.1
django-cors-headers==2.1.0
django-hosts==2.0
django-picklefield==1.0.0
django-redis-cache==1.7.1
Flask==0.12.2
Flask-Migrate==2.0.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
kombu==3.0.37
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.1
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
redis==2.10.5
six==1.10.0
SQLAlchemy==1.1.11
vine==1.1.4
virtualenv==15.1.0
virtualenvwrapper-win==1.2.1
Werkzeug==0.12.2
whitenoise==3.3.0
